Move output_file to the codegen module
